Remove reactpy.core.types
