+++ tests/tooling/common.py
@@ -1,7 +1,7 @@
 import os
 from typing import Any
 
-from reactpy.core.types import LayoutEventMessage, LayoutUpdateMessage
+from reactpy.types import LayoutEventMessage, LayoutUpdateMessage
 
 GITHUB_ACTIONS = os.getenv("GITHUB_ACTIONS", "False")
 DEFAULT_TYPE_DELAY = (
